fix order item quantity precision and active-order ownership checks

+++ services/order-service/order-service/src/main/java/com/sebet/order_service/customer/dto/response/shared/OrderItemDto.java
@@ -19,7 +19,7 @@ public record OrderItemDto(
         String productId,
         String name,
         String imageUrl,
-        int quantity,
+        BigDecimal quantity,
         BigDecimal unitPrice,
         BigDecimal subtotal
 ) {}
